# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all teams with full admin details (team lead, cadence, distribution list, member count). Requires admin (level-1) role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "List all teams (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to query teams",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new team. If a team lead is set, the supervisor chain is auto-derived from their reports-to hierarchy. Requires admin (level-1) role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "Create a team",
                "parameters": [
                    {
                        "description": "Team to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminTeamDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create team",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/teams/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a team's name, lead, cadence, and/or distribution list email. Re-derives (or clears) the supervisor chain if the team lead changes. Requires admin (level-1) role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "Update a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminTeamDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update team",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a team. Requires admin (level-1) role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "Delete a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete team",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/teams/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the members of a team. Requires admin (level-1) role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "List a team's members (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamMembersResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch team members",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a user as a member of a team. Requires admin (level-1) role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "Add a member to a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AddTeamMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found or user not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to add team member",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/teams/{id}/members/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a user from a team's membership. Requires admin (level-1) role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-teams"
                ],
                "summary": "Remove a member from a team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Team member not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to remove team member",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminTeamDTO": {
            "type": "object",
            "properties": {
                "cadence": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "distributionListEmail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "memberCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "teamLeadId": {
                    "type": "string"
                },
                "teamLeadName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.CreateTeamRequest": {
            "type": "object",
            "required": [
                "cadence",
                "name"
            ],
            "properties": {
                "cadence": {
                    "type": "string",
                    "enum": [
                        "monthly",
                        "quarterly",
                        "half-yearly",
                        "yearly"
                    ]
                },
                "distributionListEmail": {
                    "type": "string"
                },
                "id": {
                    "description": "Optional - will be auto-generated from name if not provided",
                    "type": "string"
                },
                "name": {
                    "description": "Required - used to generate ID if not provided",
                    "type": "string"
                },
                "teamLeadId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.TeamsResponse": {
            "type": "object",
            "properties": {
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.AdminTeamDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.UpdateTeamRequest": {
            "type": "object",
            "properties": {
                "cadence": {
                    "type": "string",
                    "enum": [
                        "monthly",
                        "quarterly",
                        "half-yearly",
                        "yearly"
                    ]
                },
                "distributionListEmail": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "teamLeadId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}